+++ rhine-examples/src/Periodic.hs
@@ -4,11 +4,12 @@
 -- transformers
 import Control.Monad.IO.Class
 
--- monad-schedule
-import Control.Monad.Schedule.Trans
+-- automaton
+import Data.Automaton.Schedule (YieldT, runYieldTWith)
 
 -- rhine
 import FRP.Rhine
+import Control.Concurrent (threadDelay)
 
 type MyClock = Periodic '[500, 1000]
 
@@ -17,8 +18,8 @@ everyNowAndThen =
   sinceInitS >>> proc time ->
     returnA -< unwords ["It's now", show time, "o'clock."]
 
-mainRhine :: (MonadIO m) => Rhine (ScheduleT Integer m) MyClock () ()
+mainRhine :: (MonadIO m) => Rhine (YieldT m) MyClock () ()
 mainRhine = everyNowAndThen >-> arrMCl (liftIO . putStrLn) @@ Periodic
 
 main :: IO ()
-main = runScheduleIO $ flow mainRhine
+main = runYieldTWith (threadDelay 1000) $ flow mainRhine